backend/requirements.txt
fastapi
uvicorn
sqlalchemy
sqlmodel
requests
python-dotenv
google-generativeai
pydantic
pydantic-settings
